+++ docs/module.js
@@ -1,212 +1,79 @@
-const layerTypes = {
-  'fill': ['fill-opacity'],
-  'line': ['line-opacity'],
-  'circle': ['circle-opacity', 'circle-stroke-opacity'],
-  'symbol': ['icon-opacity', 'text-opacity'],
-  'raster': ['raster-opacity'],
-  'fill-extrusion': ['fill-extrusion-opacity'],
-  'heatmap': ['heatmap-opacity']
-}
+import { YAML } from "https://js.sabae.cc/YAML.js";
+import {} from "./storytelling.js";
 
-const alignments = {
-  'left': 'lefty',
-  'center': 'centered',
-  'right': 'righty',
-  'full': 'fully'
-}
+const addStyleSheet = (href) => {
+  const link = document.createElement("link");
+  link.href = href;
+  link.rel = "stylesheet";
+  document.head.appendChild(link);
+};
+const addScript = (src) => {
+  const sc = document.createElement("script");
+  sc.src = src;
+  document.head.appendChild(sc);
+};
+const init = () => {
+  addStyleSheet("https://api.tiles.mapbox.com/mapbox-gl-js/v2.1.1/mapbox-gl.css");
+  addStyleSheet("https://optgeo.github.io/s/style.css");
+  addScript("https://api.tiles.mapbox.com/mapbox-gl-js/v2.1.1/mapbox-gl.js");
+  addScript("https://unpkg.com/intersection-observer@0.5.1/intersection-observer.js");
+  addScript("https://unpkg.com/scrollama");
 
-function getLayerPaintType(layer) {
-  const layerType = map.getLayer(layer).type;
-  return layerTypes[layerType];
-}
+  const map = document.createElement("div");
+  map.id = "map";
+  document.body.appendChild(map);
+  
+  const story = document.createElement("div");
+  story.id = "story";
+  document.body.appendChild(story);
+};
+init();
 
-function setLayerOpacity(layer) {
-  const paintProps = getLayerPaintType(layer.layer);
-  paintProps.forEach(function(prop) {
-      const options = {};
-      if (layer.duration) {
-          const transitionProp = prop + "-transition";
-          options = { "duration": layer.duration };
-          map.setPaintProperty(layer.layer, transitionProp, options);
-      }
-      map.setPaintProperty(layer.layer, prop, layer.opacity, options);
+const ALIGNMENT = 'right';
+const ROTATE = true;
+
+const process = (config) => {
+  // config.accessToken = TOKEN;
+  config.theme = 'light';
+  config.showMarkers = false;
+  let n = 0;
+  config.chapters.forEach((chapter) => {
+    n += 1;
+    chapter.id = `chapter-${n}`;
+    chapter.alignment = ALIGNMENT;
+    chapter.callback = null;
+    chapter.hidden = false;
+    chapter.mapAnimation = 'flyTo';
+    chapter.rotateAnimation = ROTATE;
+    chapter.onChapterEnter = [];
+    chapter.onChapterExit = [];
+    const r = chapter.hash.split('/');
+    console.log(n, r);
+    chapter.location = {
+      zoom: r[0],
+      center: [
+        r[2],
+        r[1]
+      ],
+      bearing: r[3],
+      pitch: r[4]
+    };
   });
-}
+  return config;
+};
 
-const story = document.getElementById('story');
-const features = document.createElement('div');
-features.setAttribute('id', 'features');
-
-const header = document.createElement('div');
-
-if (config.title) {
-  const titleText = document.createElement('h1');
-  titleText.innerText = config.title;
-  header.appendChild(titleText);
-}
-
-if (config.subtitle) {
-  const subtitleText = document.createElement('h2');
-  subtitleText.innerText = config.subtitle;
-  header.appendChild(subtitleText);
-}
-
-if (config.byline) {
-  const bylineText = document.createElement('p');
-  bylineText.innerText = config.byline;
-  header.appendChild(bylineText);
-}
-
-if (header.innerText.length > 0) {
-  header.classList.add(config.theme);
-  header.setAttribute('id', 'header');
-  story.appendChild(header);
-}
-
-config.chapters.forEach((record, idx) => {
-  const container = document.createElement('div');
-  const chapter = document.createElement('div');
-
-  if (record.title) {
-      const title = document.createElement('h3');
-      title.innerText = record.title;
-      chapter.appendChild(title);
+const getYAML = () => {
+  const sc = document.querySelectorAll("script");
+  for (const s of sc) {
+    if (s.type == "text/yaml") {
+      return s.textContent;
+    }
   }
-
-  if (record.image) {
-      const image = new Image();
-      image.src = record.image;
-      chapter.appendChild(image);
-  }
-
-  if (record.description) {
-      const story = document.createElement('p');
-      story.innerHTML = record.description;
-      chapter.appendChild(story);
-  }
-
-  container.setAttribute('id', record.id);
-  container.classList.add('step');
-  if (idx === 0) {
-      container.classList.add('active');
-  }
-
-  chapter.classList.add(config.theme);
-  container.appendChild(chapter);
-  container.classList.add(alignments[record.alignment] || 'centered');
-  if (record.hidden) {
-      container.classList.add('hidden');
-  }
-  features.appendChild(container);
-});
-
-story.appendChild(features);
-
-const footer = document.createElement('div');
-
-if (config.footer) {
-  const footerText = document.createElement('p');
-  footerText.innerHTML = config.footer;
-  footer.appendChild(footerText);
+  return null;
+};
+const yml = getYAML();
+if (!yml) {
+  alert("error: not found YAML");
+} else {
+  window.config = process(YAML.parse(yml));
 }
-
-if (footer.innerText.length > 0) {
-  footer.classList.add(config.theme);
-  footer.setAttribute('id', 'footer');
-  story.appendChild(footer);
-}
-
-mapboxgl.accessToken = config.accessToken;
-
-const transformRequest = (url) => {
-  const hasQuery = url.indexOf("?") !== -1;
-  const suffix = hasQuery ? "&pluginName=scrollytellingV2" : "?pluginName=scrollytellingV2";
-  return {
-    url: url + suffix
-  }
-}
-
-const map = new mapboxgl.Map({
-  container: 'map',
-  style: config.style,
-  center: config.chapters[0].location.center,
-  zoom: config.chapters[0].location.zoom,
-  bearing: config.chapters[0].location.bearing,
-  pitch: config.chapters[0].location.pitch,
-  interactive: false,
-  transformRequest: transformRequest
-});
-
-if (config.showMarkers) {
-  const marker = new mapboxgl.Marker({ color: config.markerColor });
-  marker.setLngLat(config.chapters[0].location.center).addTo(map);
-}
-
-// instantiate the scrollama
-const scroller = scrollama();
-
-map.on("load", function() {
-  if (config.use3dTerrain) {
-      map.addSource('mapbox-dem', {
-          'type': 'raster-dem',
-          'url': 'mapbox://mapbox.mapbox-terrain-dem-v1',
-          'tileSize': 512,
-          'maxzoom': 14
-      });
-      // add the DEM source as a terrain layer with exaggerated height
-      map.setTerrain({ 'source': 'mapbox-dem', 'exaggeration': 1.5 });
-
-      // add a sky layer that will show when the map is highly pitched
-      map.addLayer({
-          'id': 'sky',
-          'type': 'sky',
-          'paint': {
-              'sky-type': 'atmosphere',
-              'sky-atmosphere-sun': [0.0, 0.0],
-              'sky-atmosphere-sun-intensity': 15
-          }
-      });
-  };
-
-  // setup the instance, pass callback functions
-  scroller
-  .setup({
-      step: '.step',
-      offset: 0.5,
-      progress: true
-  })
-  .onStepEnter(response => {
-      const chapter = config.chapters.find(chap => chap.id === response.element.id);
-      response.element.classList.add('active');
-      map[chapter.mapAnimation || 'flyTo'](chapter.location);
-      if (config.showMarkers) {
-          marker.setLngLat(chapter.location.center);
-      }
-      if (chapter.onChapterEnter.length > 0) {
-          chapter.onChapterEnter.forEach(setLayerOpacity);
-      }
-      if (chapter.callback) {
-          chapter.callback();
-      }
-      if (chapter.rotateAnimation) {
-          map.once('moveend', function() {
-              const rotateNumber = map.getBearing();
-              map.rotateTo(rotateNumber + 90, {
-                  duration: 24000, easing: function (t) {
-                      return t;
-                  }
-              });
-          });
-      }
-  })
-  .onStepExit(response => {
-      const chapter = config.chapters.find(chap => chap.id === response.element.id);
-      response.element.classList.remove('active');
-      if (chapter.onChapterExit.length > 0) {
-          chapter.onChapterExit.forEach(setLayerOpacity);
-      }
-  });
-});
-
-// setup resize event
-window.addEventListener('resize', scroller.resize);
